Fix plink configuration option.
The third argument to AC_ARG_ENABLE is what to do if the argument is
specified and the fourth argument is what to do if the argument is not
specified. They are not true/false handlers.

Also, error out if neither ssh nor plink is available and the internal
plink was disabled.

Fixes #6.